requirements.txt
requests
beautifulsoup4
google-genai
python-dotenv
fastapi
uvicorn
